Re-organise c code and build system
