+++ microasm/src/microcode/test.h
@@ -1,8 +1,8 @@
 #ifndef TEST_H
 #define TEST_H
 
-#include "scanner.h"
-#include "parser.h"
+#include "microcode/scanner.h"
+#include "microcode/parser.h"
 
 bool runFileName(const char* fileName);
 bool runFile(const char* fileName, const char* file, Parser* parse, bool testing);
