+++ microasm/src/shared/platform.h
